requirements.txt
python-daemon
pyopenssl
django
djangorestframework
python-keyczar
django_extensions